pyproject.toml
[project]
name = "assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.2.2",
    "fastapi>=0.120.1",
    "openai-agents>=0.4.2",
    "pypdf>=6.1.3",
    "python-pptx>=1.0.2",
    "motor>=3.6.0",
    "pymongo>=4.10.1",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.20",
    "pydantic[email]>=2.10.5",
    "uvicorn>=0.34.0",
]
